// src/modules/featureConfiguration/enum/featureConfigurationFeatureConfig.enum.ts
export enum FeatureConfigKeyName {
  Id = 'id',
  CaseCategory = 'caseCategory',
  CreatedAt = 'createdAt',
  Location = 'location',
  Subject = 'subject',
  Status = 'status',
  UpdatedAt = 'updatedAt',
  NameAndImage = 'nameAndImage',
  Name = 'name',
  ShortId = 'shortId',
  DateStart = 'dateStart',
  DateEnd = 'dateEnd',
  Type = 'type',
  AssociateName = 'associateName',
  Service = 'service',
  LegalEntity = 'legalEntity',
  ServiceType = 'serviceType',
  Number = 'number',
  SignatureDate = 'signatureDate',
  SignedBy = 'signedBy',
  HoursBilled = 'hoursBilled',
  TotalAmount = 'totalAmount',
  DuePaymentDate = 'duePaymentDate',
  Overdue = 'overdue',
  IssueDate = 'issueDate',
  PeriodStart = 'periodStart',
  PeriodEnd = 'periodEnd',
  CreditNotes = 'creditNotes',
  AmountBeforeTax = 'amountBeforeTax',
  Currency = 'currency',
  Attachments = 'attachments',
  Title = 'title',
  UserType = 'userType',
  Worksite = 'worksite',
  Permissions = 'permissions',
  Department = 'department',
  Function = 'function',
  MainLocation = 'mainLocation',
  StartAndEndDate = 'startAndEndDate',
  NoOfPositions = 'noOfPositions',
  Role = 'role',
  Shift = 'shift',
}

export enum FeatureConfigKeyType {
  String = 'string',
  Number = 'number',
  CustomObject = 'customObject',
  Date = 'date',
  Status = 'status',
  InputFilter = 'input-filter',
  SingleCheckbox = 'singleCheckbox',
  Link = 'link',
  Array = 'array',
  NumberOfPositions = 'numberOfPositions',
}

export enum FeatureConfigUserRole {
  Admin = 'admin',
  Staff = 'staff',
}